refacto split error.c file into a firmware/error.c and an emulator/error.c files ref #226
